fix bugs in images

+++ src/services/images.py
@@ -1,12 +1,19 @@
+import logging
+import os
 import shutil
 
 from src.services.base import BaseService
 from src.tasks.task import resize_image
+from src.config import settings
 
+STATIC_DIR = "/home/eugen/Python_Projects/Hotels/src/static/images"
 
 class ImageService(BaseService):
     def upload_image(self, image) -> None:
-        image_path = f"Hotels/src/static/images/{image.filename}"
+        image_path = os.path.join(settings.STATIC_DIR, image.filename)
+        if os.path.exists(image_path):
+            raise FileExistsError(f"Файл {image.filename} уже существует")
+        logging.info(f"Обработка изображения: {image_path}")
         try:
             with open(image_path, "wb+") as f:
                 shutil.copyfileobj(image.file, f)
